use size_t for buffer indices/capacity; simplify unsigned literals in tests
Replace size_t{N} with NU in test assertions for clarity. Update
CircularBuffer to use size_t for capacity and index types. Use
unsigned int instead of size_t for string substring counts.

+++ examples/include/CircularBuffer.hpp
@@ -1,19 +1,14 @@
 #ifndef INCLUDED_CIRCULARBUFFER_HPP
 #define INCLUDED_CIRCULARBUFFER_HPP
 
-///////////////////////////////////////////////////////////////////////////////
-//
-//  CircularBuffer.h
-//
-//  CircularBuffer is responsible for ...
-//
-///////////////////////////////////////////////////////////////////////////////
+#include <stddef.h>
+
 class Printer;
 
 class CircularBuffer
 {
 public:
-  explicit CircularBuffer(int capacity = default_capacity);
+  explicit CircularBuffer(size_t capacity = default_capacity);
   virtual ~CircularBuffer();
 
   CircularBuffer(const CircularBuffer&) = delete;
@@ -23,16 +18,16 @@ public:
   int get();
   bool is_empty() const;
   bool is_full() const;
-  int capacity() const;
-  int next(int i) const;
+  size_t capacity() const;
+  size_t next(size_t i) const;
   void print(Printer* p);
 
 private:
-  int index_{ 0 };
-  int outdex_{ 0 };
+  size_t index_{ 0 };
+  size_t outdex_{ 0 };
   int* buffer_;
-  int capacity_;
-  static constexpr int default_capacity = 5;
+  size_t capacity_;
+  static constexpr size_t default_capacity = 5;
   bool empty_{ true };
   bool full_{ false };
 };
